Added time/block switcher
