+++ app/components/ScheduleWizard/TimeSettings/index.js
@@ -1,0 +1,2 @@
+export { TimeComponent } from './TimeComponent'
+export { BlockComponent } from './BlockComponent'
